inntektsmelding apitest, organisering api-tester
