+++ src/test/kotlin/no/nav/helsearbeidsgiver/authorization/HentForespoerslerUnauthorizedApiTest.kt
@@ -1,4 +1,4 @@
-package no.nav.helsearbeidsgiver.forespoersel
+package no.nav.helsearbeidsgiver.authorization
 
 import io.kotest.matchers.shouldBe
 import no.nav.helsearbeidsgiver.utils.FunSpecWithUnauthorizedApi
@@ -7,8 +7,10 @@ class HentForespoerslerUnauthorizedApiTest :
     FunSpecWithUnauthorizedApi(emptyMap(), { testApi ->
         test("Gir 401 når claim mangler i token") {
             testApi {
-                val response = get("/forespoersler")
-                response.status.value shouldBe 401
+                val response1 = get("/forespoersler")
+                response1.status.value shouldBe 401
+                val response2 = get("/inntektsmeldinger")
+                response2.status.value shouldBe 401
             }
         }
     })
